[retroplayer] - (0006.5) Expose stream player info to GUI
Follow-up to PR 9976, "VideoPlayer: expose stream player info to GUI".

This PR gets a .5 index because it is dependent on the commit
"DVDVideoCodecFFmpeg: Expose function to get pixel format name".

That commit is probably not the right way to access ffmpeg's pixel format
names. When this problem is fixed, (0006.5) can be squashed into (0006).