Game API v4.0.0: Add hardware rendering support (WIP)
This change adds a new callback to enable hardware rendering.

Previously, we relied on lazy enabling of hardware rendering inside
OpenStream(). This doesn't work, as the OpenGL parameters are needed at
an earlier stage.

Now, a new function, EnableHardwareRendering(), is added to accept these
parameters at an earlier time. An empty struct is then passed to
OpenStream().