[PR 120] Remove use of default parameter
Default parameters should be avoided, because it adds additional overhead
to understandability of the code. In this case, autosave or not is a binary
decision, and it's not clear why "automatic" should logically be the default.
Therefore, there's no reason to assume the reader will know that a missing
parameter means "automatic".