[PR 18418] [addons] add reworked serialization way for addon extensions on database
This changes the way the field of the addon.xml of `<extensions point="addon_type_name">...</extension>` is stored in the database.

Previously, only the addon type string was stored as "extension" in json and any sub-data was stored separately under "extrainfo". This makes it difficult to process more complex data.

This introduces a way to save the complete extension's in a json format and when reading all the data of the main addon type is available.

**WARNING:**
A way has to be introduced in order to update the add-on database appropriately, otherwise the content of the previous version is no longer visible in Kodi.
As a temporary test way, the old "Addons32.db" can be deleted so that it is recreated.