[PR 21713] Controllers: Fix segfault when features can't be created
During development, some feature definitions were missing, causing features
to not be created. However, we pre-allocate features in the array, which
means that until the feature is created, the array holds a nullptr.

Later, we use the array without checking for nullptrs, leading to a segfault.
Instead of checking for nullptrs, though, we just ensure that no nullptrs
enter the array in the first place.