[PR 21585] [cmake] enable copy_file_to_buildtree to handle files created at build time
This allows files in DEPENDS_PATH to be handled at build time to install to
the target build tree. This is required when for example a shared lib is built at build time
and needs to be bundled. The existing globbing wont find the lib at generation time,
therefore it doesnt get bundled.