[PR 25011] [video] Fix Movie Information Refresh
For the case of information dialog opened on folder named after a movie and containing single movie

Better fix would be at a lower level (for example CFileItem::GetBaseMoviePath()) but complicated
at this time because of videos with versions as folder. They wouldn't be handled correctly anymore.
They look like folders and at a lower level lack information to distinguish from actual folders.